@@ -1,6 +1,7 @@
 package simapp
 
 import (
+	"encoding/json"
 	"time"
 
 	"github.com/cosmos/cosmos-sdk/simapp"
@@ -21,12 +22,15 @@ func New(dir string) *app.App {
 	encoding := app.MakeEncodingConfig()
 
 	a := app.NewOmniFlixApp(logger, db, nil, true, map[int64]bool{}, dir, 0, encoding,
-		// this line is used by starport scaffolding # stargate/testutil/appArgument
 		simapp.EmptyAppOptions{})
+	stateBytes, err := json.MarshalIndent(app.ModuleBasics.DefaultGenesis(encoding.Marshaler), "", " ")
+	if err != nil {
+		panic(err)
+	}
 	// InitChain updates deliverState which is required when app.NewContext is called
 	a.InitChain(abci.RequestInitChain{
 		ConsensusParams: defaultConsensusParams,
-		AppStateBytes:   []byte("{}"),
+		AppStateBytes:   stateBytes,
 	})
 	return a
 }
