@@ -32,7 +32,10 @@ func NewKeeper(
 	storeKey,
 	memKey sdk.StoreKey,
 
-	accountKeeper types.AccountKeeper, bankKeeper types.BankKeeper, stakingKeeper types.StakingKeeper, distrKeeper types.DistrKeeper,
+	accountKeeper types.AccountKeeper,
+	bankKeeper types.BankKeeper,
+	stakingKeeper types.StakingKeeper,
+	distrKeeper types.DistrKeeper,
 	ps paramtypes.Subspace,
 ) *Keeper {
 	// set KeyTable if it has not already been set
@@ -57,8 +60,8 @@ func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 	return ctx.Logger().With("module", fmt.Sprintf("x/%s", types.ModuleName))
 }
 
-// GetModuleAccountBalance gets the coin balance of module account
-func (k Keeper) GetModuleAccountAddress(ctx sdk.Context) sdk.AccAddress {
+// GetModuleAccountAddress gets the address of module account
+func (k Keeper) GetModuleAccountAddress() sdk.AccAddress {
 	return k.accountKeeper.GetModuleAddress(types.ModuleName)
 }
 
@@ -66,66 +69,104 @@ func (k Keeper) GetModuleAccountAddress(ctx sdk.Context) sdk.AccAddress {
 func (k Keeper) DistributeMintedCoins(ctx sdk.Context) error {
 	blockRewardsAddr := k.accountKeeper.GetModuleAccount(ctx, authtypes.FeeCollectorName).GetAddress()
 	blockRewards := k.bankKeeper.GetBalance(ctx, blockRewardsAddr, k.stakingKeeper.BondDenom(ctx))
-	blockRewardsDec := sdk.NewDecFromInt(blockRewards.Amount)
+	blockRewardsAmountDec := blockRewards.Amount.ToDec()
 
 	params := k.GetParams(ctx)
 	proportions := params.DistributionProportions
 
-	nftIncentiveAmount := blockRewardsDec.Mul(proportions.NftIncentives)
-	nftIncentiveCoin := sdk.NewCoin(k.stakingKeeper.BondDenom(ctx), nftIncentiveAmount.TruncateInt())
-	// Distribute NFT incentives to the community pool until a future update
-	err := k.distrKeeper.FundCommunityPool(ctx, sdk.NewCoins(nftIncentiveCoin), blockRewardsAddr)
-	if err != nil {
+	nftIncentiveAmount := blockRewardsAmountDec.Mul(proportions.NftIncentives).TruncateInt()
+	nftIncentiveCoin := sdk.NewCoin(k.stakingKeeper.BondDenom(ctx), nftIncentiveAmount)
+
+	k.Logger(ctx).Debug(
+		"distributing minted coins to nft incentives receivers",
+		"amount", nftIncentiveCoin.String(), "from", blockRewardsAddr,
+	)
+	if err := k.distributeCoinToWeightedAddresses(
+		ctx,
+		params.WeightedNftIncentivesReceivers,
+		nftIncentiveCoin,
+		blockRewardsAddr,
+	); err != nil {
 		return err
 	}
-	k.Logger(ctx).Debug("funded community pool with nft incentives", "amount", nftIncentiveCoin.String(), "from", blockRewardsAddr)
 
-	nodeHostsIncentiveAmount := blockRewardsDec.Mul(proportions.NodeHostsIncentives)
-	nodeHostsIncentiveCoin := sdk.NewCoin(k.stakingKeeper.BondDenom(ctx), nodeHostsIncentiveAmount.TruncateInt())
-	// Distribute node hosts incentives to the community pool until a future update
-	err = k.distrKeeper.FundCommunityPool(ctx, sdk.NewCoins(nodeHostsIncentiveCoin), blockRewardsAddr)
-	if err != nil {
+	nodeHostsIncentiveAmount := blockRewardsAmountDec.Mul(proportions.NodeHostsIncentives).TruncateInt()
+	nodeHostsIncentiveCoin := sdk.NewCoin(k.stakingKeeper.BondDenom(ctx), nodeHostsIncentiveAmount)
+
+	k.Logger(ctx).Debug(
+		"distributing minted coins to node host incentives receivers",
+		"amount", nodeHostsIncentiveCoin.String(), "from", blockRewardsAddr,
+	)
+	if err := k.distributeCoinToWeightedAddresses(
+		ctx,
+		params.WeightedNodeHostsIncentivesReceivers,
+		nodeHostsIncentiveCoin,
+		blockRewardsAddr,
+	); err != nil {
 		return err
 	}
-	k.Logger(ctx).Debug("funded community pool with node host incentives", "amount", nodeHostsIncentiveCoin.String(), "from", blockRewardsAddr)
 
-	devRewardAmount := blockRewardsDec.Mul(proportions.DeveloperRewards)
-	devRewardCoin := sdk.NewCoin(k.stakingKeeper.BondDenom(ctx), devRewardAmount.TruncateInt())
+	devRewardAmount := blockRewardsAmountDec.Mul(proportions.DeveloperRewards).TruncateInt()
+	devRewardCoin := sdk.NewCoin(k.stakingKeeper.BondDenom(ctx), devRewardAmount)
 
-	for _, w := range params.WeightedDeveloperRewardsReceivers {
-		devRewardPortionCoins := sdk.NewCoins(k.GetProportions(ctx, devRewardCoin, w.Weight))
-		if w.Address == "" {
-			err := k.distrKeeper.FundCommunityPool(ctx, devRewardPortionCoins, blockRewardsAddr)
-			if err != nil {
-				return err
-			}
-		} else {
-			devRewardsAddr, err := sdk.AccAddressFromBech32(w.Address)
-			if err != nil {
-				return err
-			}
-			err = k.bankKeeper.SendCoins(ctx, blockRewardsAddr, devRewardsAddr, devRewardPortionCoins)
-			if err != nil {
-				return err
-			}
-			k.Logger(ctx).Debug("sent coins to developer", "amount", devRewardPortionCoins.String(), "from", blockRewardsAddr, "to", devRewardsAddr)
-		}
+	k.Logger(ctx).Debug(
+		"distributing minted coins to developer rewards receivers",
+		"amount", devRewardCoin.String(), "from", blockRewardsAddr,
+	)
+
+	if err := k.distributeCoinToWeightedAddresses(
+		ctx,
+		params.WeightedDeveloperRewardsReceivers,
+		devRewardCoin,
+		blockRewardsAddr,
+	); err != nil {
+		return err
 	}
+
 	// calculate staking rewards
-	stakingRewardsCoins := sdk.NewCoins(k.GetProportions(ctx, blockRewards, proportions.StakingRewards))
+	stakingRewardAmount := blockRewardsAmountDec.Mul(proportions.StakingRewards).TruncateInt()
+	stakingRewardCoin := sdk.NewCoin(k.stakingKeeper.BondDenom(ctx), stakingRewardAmount)
 
 	// subtract from original provision to ensure no coins left over after the allocations
-	communityPoolCoins := sdk.NewCoins(blockRewards).Sub(stakingRewardsCoins).Sub(sdk.NewCoins(nftIncentiveCoin)).Sub(sdk.NewCoins(nodeHostsIncentiveCoin)).Sub(sdk.NewCoins(devRewardCoin))
-	err = k.distrKeeper.FundCommunityPool(ctx, communityPoolCoins, blockRewardsAddr)
-	if err != nil {
+	communityPoolCoin := blockRewards.
+		Sub(stakingRewardCoin).
+		Sub(nftIncentiveCoin).
+		Sub(nodeHostsIncentiveCoin).
+		Sub(devRewardCoin)
+
+	if err := k.distrKeeper.FundCommunityPool(ctx, sdk.NewCoins(communityPoolCoin), blockRewardsAddr); err != nil {
 		return err
 	}
 
 	return nil
 }
 
-// GetProportions gets the balance of the `MintedDenom` from minted coins
-// and returns coins according to the `AllocationRatio`
-func (k Keeper) GetProportions(ctx sdk.Context, mintedCoin sdk.Coin, ratio sdk.Dec) sdk.Coin {
-	return sdk.NewCoin(mintedCoin.Denom, mintedCoin.Amount.ToDec().Mul(ratio).TruncateInt())
+func (k Keeper) distributeCoinToWeightedAddresses(
+	ctx sdk.Context,
+	weightedAddresses []types.WeightedAddress,
+	totalAmount sdk.Coin,
+	fromAddress sdk.AccAddress,
+) error {
+	totalAmountDec := totalAmount.Amount.ToDec()
+	for _, w := range weightedAddresses {
+		amount := sdk.NewCoin(k.stakingKeeper.BondDenom(ctx), totalAmountDec.Mul(w.Weight).TruncateInt())
+		if w.Address == "" {
+			err := k.distrKeeper.FundCommunityPool(ctx, sdk.NewCoins(amount), fromAddress)
+			if err != nil {
+				return err
+			}
+		} else {
+			toAddress, err := sdk.AccAddressFromBech32(w.Address)
+			if err != nil {
+				return err
+			}
+			err = k.bankKeeper.SendCoins(ctx, fromAddress, toAddress, sdk.NewCoins(amount))
+			if err != nil {
+				return err
+			}
+			k.Logger(ctx).Debug(
+				"sent coins to address", "amount", amount.String(), "from", fromAddress, "to", toAddress)
+		}
+	}
+	return nil
 }
