@@ -4,10 +4,7 @@ import (
 	"encoding/json"
 
 	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 )
-
-// this line is used by starport scaffolding # genesis/types/import
 
 // DefaultIndex is the default capability global index
 const DefaultIndex uint64 = 1
@@ -15,16 +12,7 @@ const DefaultIndex uint64 = 1
 // DefaultGenesis returns the default Capability genesis state
 func DefaultGenesis() *GenesisState {
 	return &GenesisState{
-		Params: Params{
-			DistributionProportions: DistributionProportions{
-				StakingRewards:      sdk.NewDecWithPrec(35, 2), // 35%
-				NftIncentives:       sdk.NewDecWithPrec(40, 2), // 40%
-				NodeHostsIncentives: sdk.NewDecWithPrec(5, 2),  // 5%
-				DeveloperRewards:    sdk.NewDecWithPrec(15, 2), // 15%
-				CommunityPool:       sdk.NewDecWithPrec(5, 2),  // 5%
-			},
-			WeightedDeveloperRewardsReceivers: []WeightedAddress{},
-		},
+		Params: DefaultParams(),
 	}
 }
 
