@@ -114,7 +114,11 @@ type Params struct {
 	// distribution_proportions defines the proportion of the minted denom
 	DistributionProportions DistributionProportions `protobuf:"bytes,1,opt,name=distribution_proportions,json=distributionProportions,proto3" json:"distribution_proportions"`
 	// address to receive developer rewards
-	WeightedDeveloperRewardsReceivers []WeightedAddress `protobuf:"bytes,2,rep,name=weighted_developer_rewards_receivers,json=weightedDeveloperRewardsReceivers,proto3" json:"weighted_developer_rewards_receivers" yaml:"developer_rewards_receivers"`
+	WeightedDeveloperRewardsReceivers []WeightedAddress `protobuf:"bytes,2,rep,name=weighted_developer_rewards_receivers,json=weightedDeveloperRewardsReceivers,proto3" json:"weighted_developer_rewards_receivers" yaml:"weighted_developer_rewards_receivers"`
+	// address to receive nft incentives
+	WeightedNftIncentivesReceivers []WeightedAddress `protobuf:"bytes,3,rep,name=weighted_nft_incentives_receivers,json=weightedNftIncentivesReceivers,proto3" json:"weighted_nft_incentives_receivers" yaml:"weighted_nft_incentives_receivers"`
+	// address to receive node host incentives
+	WeightedNodeHostsIncentivesReceivers []WeightedAddress `protobuf:"bytes,4,rep,name=weighted_node_hosts_incentives_receivers,json=weightedNodeHostsIncentivesReceivers,proto3" json:"weighted_node_hosts_incentives_receivers" yaml:"weighted_node_hosts_incentives_receivers"`
 }
 
 func (m *Params) Reset()         { *m = Params{} }
@@ -164,6 +168,20 @@ func (m *Params) GetWeightedDeveloperRewardsReceivers() []WeightedAddress {
 	return nil
 }
 
+func (m *Params) GetWeightedNftIncentivesReceivers() []WeightedAddress {
+	if m != nil {
+		return m.WeightedNftIncentivesReceivers
+	}
+	return nil
+}
+
+func (m *Params) GetWeightedNodeHostsIncentivesReceivers() []WeightedAddress {
+	if m != nil {
+		return m.WeightedNodeHostsIncentivesReceivers
+	}
+	return nil
+}
+
 func init() {
 	proto.RegisterType((*WeightedAddress)(nil), "omniflix.alloc.v1beta1.WeightedAddress")
 	proto.RegisterType((*DistributionProportions)(nil), "omniflix.alloc.v1beta1.DistributionProportions")
@@ -175,41 +193,45 @@ func init() {
 }
 
 var fileDescriptor_e15b25018432ff35 = []byte{
-	// 542 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x94, 0xbf, 0x6f, 0xd3, 0x40,
-	0x14, 0xc7, 0xe3, 0xb6, 0x04, 0x71, 0x55, 0x52, 0x30, 0xb4, 0x8d, 0x10, 0xb2, 0xe1, 0x40, 0x50,
-	0x21, 0x6a, 0xab, 0x65, 0x63, 0x41, 0x44, 0x11, 0x2d, 0x0c, 0x10, 0xdd, 0x52, 0x89, 0xc5, 0xf2,
-	0x8f, 0x8b, 0x73, 0xaa, 0x7d, 0x67, 0xee, 0x2e, 0x09, 0x59, 0xf9, 0x0b, 0x18, 0x59, 0x91, 0xf8,
-	0x63, 0x3a, 0x76, 0x44, 0x0c, 0x16, 0x4a, 0x36, 0xc6, 0xfc, 0x05, 0xc8, 0x67, 0x3b, 0x75, 0xd3,
-	0x66, 0x88, 0x98, 0x72, 0xf7, 0xf2, 0xd5, 0xf7, 0xf3, 0xfc, 0xee, 0xbd, 0x07, 0x1e, 0xb3, 0x98,
-	0x92, 0x5e, 0x44, 0xbe, 0xd8, 0x6e, 0x14, 0x31, 0xdf, 0x1e, 0x1e, 0x78, 0x58, 0xba, 0x07, 0x76,
-	0xe2, 0x72, 0x37, 0x16, 0x56, 0xc2, 0x99, 0x64, 0xfa, 0x4e, 0x29, 0xb2, 0x94, 0xc8, 0x2a, 0x44,
-	0xf7, 0xef, 0x85, 0x2c, 0x64, 0x4a, 0x62, 0x67, 0xa7, 0x5c, 0x0d, 0xbf, 0x6b, 0x60, 0xeb, 0x04,
-	0x93, 0xb0, 0x2f, 0x71, 0xf0, 0x26, 0x08, 0x38, 0x16, 0x42, 0x7f, 0x01, 0x6e, 0xba, 0xf9, 0xb1,
-	0xa5, 0x3d, 0xd4, 0xf6, 0x6e, 0xb5, 0xf5, 0x59, 0x6a, 0x36, 0xc7, 0x6e, 0x1c, 0xbd, 0x82, 0xc5,
-	0x1f, 0x10, 0x95, 0x12, 0xfd, 0x04, 0xd4, 0x47, 0xca, 0xa0, 0xb5, 0xa6, 0xc4, 0xaf, 0xcf, 0x52,
-	0xb3, 0xf6, 0x3b, 0x35, 0x9f, 0x86, 0x44, 0xf6, 0x07, 0x9e, 0xe5, 0xb3, 0xd8, 0xf6, 0x99, 0x88,
-	0x99, 0x28, 0x7e, 0xf6, 0x45, 0x70, 0x6a, 0xcb, 0x71, 0x82, 0x85, 0xd5, 0xc1, 0xfe, 0x2c, 0x35,
-	0x1b, 0xb9, 0x75, 0xee, 0x02, 0x51, 0x61, 0x07, 0xff, 0x6e, 0x80, 0xdd, 0x0e, 0x11, 0x92, 0x13,
-	0x6f, 0x20, 0x09, 0xa3, 0x5d, 0xce, 0x12, 0xc6, 0xb3, 0x93, 0xd0, 0x3f, 0x83, 0x2d, 0x21, 0xdd,
-	0x53, 0x42, 0x43, 0x87, 0xe3, 0x91, 0xcb, 0x83, 0x32, 0xd5, 0xe3, 0x95, 0xe9, 0x3b, 0x39, 0x7d,
-	0xc1, 0x0e, 0xa2, 0x66, 0x11, 0x41, 0x79, 0x40, 0xa7, 0xa0, 0x49, 0x7b, 0xd2, 0x21, 0xd4, 0xc7,
-	0x54, 0x92, 0x21, 0x16, 0xc5, 0xf7, 0x1e, 0xad, 0x4c, 0xdc, 0xce, 0x89, 0x97, 0xdd, 0x20, 0x6a,
-	0xd0, 0x9e, 0x7c, 0x37, 0xbf, 0xeb, 0x5f, 0x35, 0xb0, 0x4d, 0x59, 0x80, 0x9d, 0x3e, 0x13, 0x52,
-	0x54, 0xb9, 0xeb, 0x8a, 0xfb, 0x61, 0x65, 0xee, 0x83, 0x82, 0x7b, 0x9d, 0x29, 0x44, 0x77, 0xb3,
-	0xf8, 0x71, 0x16, 0xae, 0x24, 0x31, 0x02, 0x77, 0x02, 0x3c, 0xc4, 0x11, 0x4b, 0x30, 0x9f, 0x57,
-	0x7a, 0x43, 0xf1, 0xdf, 0xaf, 0xcc, 0x6f, 0xe5, 0xfc, 0x2b, 0x86, 0x10, 0xdd, 0x9e, 0xc7, 0x2a,
-	0xd5, 0xf6, 0x59, 0x1c, 0x0f, 0x28, 0x91, 0x63, 0x27, 0x61, 0x2c, 0x6a, 0xdd, 0xf8, 0xbf, 0x6a,
-	0x5f, 0x76, 0x83, 0xa8, 0x31, 0x0f, 0x74, 0xb3, 0xfb, 0x8f, 0x35, 0x50, 0xef, 0xaa, 0x31, 0xd2,
-	0x13, 0xd0, 0x0a, 0x2a, 0x6d, 0xe7, 0x24, 0x17, 0x7d, 0xa7, 0x9a, 0x6c, 0xf3, 0xd0, 0xb6, 0xae,
-	0x9f, 0x31, 0x6b, 0x49, 0xbb, 0xb6, 0x37, 0xb2, 0xac, 0xd1, 0x6e, 0xb0, 0xa4, 0x9b, 0x7f, 0x6a,
-	0xe0, 0xc9, 0xa8, 0x18, 0x42, 0xe7, 0x4a, 0x79, 0x1c, 0x8e, 0x7d, 0x4c, 0x86, 0x98, 0x67, 0x1d,
-	0xb7, 0xbe, 0xb7, 0x79, 0xf8, 0x6c, 0x19, 0x7e, 0x61, 0x90, 0xdb, 0xcf, 0x33, 0xec, 0x2c, 0x35,
-	0xe1, 0x92, 0xc2, 0x5f, 0x38, 0x43, 0xf4, 0xa8, 0x4c, 0xa0, 0xb3, 0xf0, 0x14, 0xa8, 0xd4, 0xb4,
-	0x8f, 0xce, 0x26, 0x86, 0x76, 0x3e, 0x31, 0xb4, 0x3f, 0x13, 0x43, 0xfb, 0x36, 0x35, 0x6a, 0xe7,
-	0x53, 0xa3, 0xf6, 0x6b, 0x6a, 0xd4, 0x3e, 0xed, 0x57, 0x5e, 0xe3, 0x63, 0x4c, 0xc9, 0xdb, 0x6c,
-	0x47, 0x95, 0x49, 0xf6, 0x07, 0x9e, 0x5d, 0x6e, 0x2c, 0xf5, 0x30, 0x5e, 0x5d, 0xed, 0x9e, 0x97,
-	0xff, 0x02, 0x00, 0x00, 0xff, 0xff, 0xd6, 0x97, 0xdb, 0xeb, 0xd0, 0x04, 0x00, 0x00,
+	// 605 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x55, 0x3f, 0x6f, 0xd3, 0x40,
+	0x1c, 0x8d, 0x49, 0x1a, 0xc4, 0x55, 0x49, 0xc1, 0xd0, 0x36, 0x42, 0xc8, 0x81, 0x03, 0x41, 0x24,
+	0xa8, 0x4d, 0xdb, 0x01, 0x89, 0x05, 0x11, 0x45, 0xb4, 0x30, 0x84, 0xe8, 0x96, 0x4a, 0x2c, 0x96,
+	0x63, 0x5f, 0x92, 0x53, 0x6d, 0x9f, 0xb9, 0xbb, 0x24, 0x64, 0x65, 0x66, 0x60, 0xe4, 0x03, 0xb0,
+	0xb2, 0xf3, 0x11, 0x3a, 0x76, 0x44, 0x0c, 0x11, 0x4a, 0x36, 0xc6, 0x7c, 0x02, 0xe4, 0xbf, 0x71,
+	0xfe, 0x55, 0xb2, 0x98, 0x72, 0xfe, 0xe5, 0xe9, 0xbd, 0x97, 0x9f, 0xdf, 0xcb, 0x81, 0x87, 0xd4,
+	0x71, 0x49, 0xc7, 0x26, 0x9f, 0x34, 0xc3, 0xb6, 0xa9, 0xa9, 0x0d, 0x0e, 0xdb, 0x58, 0x18, 0x87,
+	0x9a, 0x67, 0x30, 0xc3, 0xe1, 0xaa, 0xc7, 0xa8, 0xa0, 0xf2, 0x5e, 0x0c, 0x52, 0x03, 0x90, 0x1a,
+	0x81, 0xee, 0xde, 0xe9, 0xd2, 0x2e, 0x0d, 0x20, 0x9a, 0x7f, 0x0a, 0xd1, 0xf0, 0x9b, 0x04, 0x76,
+	0xce, 0x30, 0xe9, 0xf6, 0x04, 0xb6, 0x5e, 0x5b, 0x16, 0xc3, 0x9c, 0xcb, 0xcf, 0xc0, 0x75, 0x23,
+	0x3c, 0x56, 0xa4, 0xfb, 0x52, 0xed, 0x46, 0x5d, 0x9e, 0x8d, 0xab, 0xe5, 0x91, 0xe1, 0xd8, 0x2f,
+	0x61, 0xf4, 0x05, 0x44, 0x31, 0x44, 0x3e, 0x03, 0xc5, 0x61, 0x40, 0x50, 0xb9, 0x16, 0x80, 0x5f,
+	0x5d, 0x8c, 0xab, 0xb9, 0xdf, 0xe3, 0xea, 0xe3, 0x2e, 0x11, 0xbd, 0x7e, 0x5b, 0x35, 0xa9, 0xa3,
+	0x99, 0x94, 0x3b, 0x94, 0x47, 0x1f, 0x07, 0xdc, 0x3a, 0xd7, 0xc4, 0xc8, 0xc3, 0x5c, 0x6d, 0x60,
+	0x73, 0x36, 0xae, 0x96, 0x42, 0xea, 0x90, 0x05, 0xa2, 0x88, 0x0e, 0xfe, 0x2d, 0x80, 0xfd, 0x06,
+	0xe1, 0x82, 0x91, 0x76, 0x5f, 0x10, 0xea, 0xb6, 0x18, 0xf5, 0x28, 0xf3, 0x4f, 0x5c, 0xfe, 0x08,
+	0x76, 0xb8, 0x30, 0xce, 0x89, 0xdb, 0xd5, 0x19, 0x1e, 0x1a, 0xcc, 0x8a, 0xad, 0x9e, 0x66, 0x56,
+	0xdf, 0x0b, 0xd5, 0x97, 0xe8, 0x20, 0x2a, 0x47, 0x13, 0x14, 0x0e, 0x64, 0x17, 0x94, 0xdd, 0x8e,
+	0xd0, 0x89, 0x6b, 0x62, 0x57, 0x90, 0x01, 0xe6, 0xd1, 0xef, 0x3d, 0xc9, 0xac, 0xb8, 0x1b, 0x2a,
+	0x2e, 0xb2, 0x41, 0x54, 0x72, 0x3b, 0xe2, 0x6d, 0xf2, 0x2c, 0x7f, 0x96, 0xc0, 0xae, 0x4b, 0x2d,
+	0xac, 0xf7, 0x28, 0x17, 0x3c, 0xad, 0x9b, 0x0f, 0x74, 0x9b, 0x99, 0x75, 0xef, 0x45, 0xba, 0xeb,
+	0x48, 0x21, 0xba, 0xed, 0xcf, 0x4f, 0xfd, 0x71, 0xca, 0xc4, 0x10, 0xdc, 0xb2, 0xf0, 0x00, 0xdb,
+	0xd4, 0xc3, 0x2c, 0xd9, 0x74, 0x21, 0xd0, 0x7f, 0x97, 0x59, 0xbf, 0x12, 0xea, 0xaf, 0x10, 0x42,
+	0x74, 0x33, 0x99, 0xa5, 0xb6, 0x6d, 0x52, 0xc7, 0xe9, 0xbb, 0x44, 0x8c, 0x74, 0x8f, 0x52, 0xbb,
+	0xb2, 0xf5, 0x7f, 0xdb, 0x5e, 0x64, 0x83, 0xa8, 0x94, 0x0c, 0x5a, 0xfe, 0xf3, 0x97, 0x2d, 0x50,
+	0x6c, 0x05, 0x35, 0x92, 0x3d, 0x50, 0xb1, 0x52, 0xb1, 0xd3, 0xbd, 0x79, 0xee, 0x82, 0x90, 0x6d,
+	0x1f, 0x69, 0xea, 0xfa, 0x8e, 0xa9, 0x1b, 0xe2, 0x5a, 0x2f, 0xf8, 0xae, 0xd1, 0xbe, 0xb5, 0x21,
+	0xcd, 0x3f, 0x24, 0xf0, 0x68, 0x18, 0x95, 0x50, 0x5f, 0x59, 0x8f, 0xce, 0xb0, 0x89, 0xc9, 0x00,
+	0x33, 0x3f, 0x71, 0xf9, 0xda, 0xf6, 0xd1, 0x93, 0x4d, 0xf2, 0x4b, 0x45, 0xae, 0x1f, 0xfb, 0xb2,
+	0xb3, 0x71, 0xf5, 0x69, 0xba, 0x60, 0x57, 0x4b, 0x40, 0xf4, 0x20, 0x86, 0x35, 0x96, 0xde, 0x09,
+	0x8a, 0x31, 0xf2, 0x77, 0x09, 0x24, 0x28, 0x7d, 0x31, 0xc6, 0x29, 0xb3, 0xf9, 0x6c, 0x66, 0x9f,
+	0x47, 0x66, 0x6b, 0x4b, 0x66, 0x37, 0xf1, 0x43, 0xa4, 0xc4, 0x98, 0x66, 0xba, 0x38, 0x73, 0x9b,
+	0x3f, 0x25, 0x50, 0x9b, 0xd3, 0xac, 0x4b, 0x7d, 0xca, 0x6d, 0x21, 0x9b, 0xdb, 0x17, 0x91, 0x5b,
+	0x6d, 0xd9, 0xed, 0xd5, 0x32, 0x10, 0x25, 0x2f, 0xba, 0xb9, 0x5a, 0xb7, 0xc4, 0x7a, 0xfd, 0xe4,
+	0x62, 0xa2, 0x48, 0x97, 0x13, 0x45, 0xfa, 0x33, 0x51, 0xa4, 0xaf, 0x53, 0x25, 0x77, 0x39, 0x55,
+	0x72, 0xbf, 0xa6, 0x4a, 0xee, 0xc3, 0x41, 0x2a, 0xf8, 0xef, 0x1d, 0x97, 0xbc, 0xf1, 0xaf, 0x83,
+	0xd8, 0x74, 0xaf, 0xdf, 0xd6, 0xe2, 0xcb, 0x21, 0xe8, 0x40, 0xbb, 0x18, 0xfc, 0xcd, 0x1f, 0xff,
+	0x0b, 0x00, 0x00, 0xff, 0xff, 0x04, 0xd1, 0xae, 0xc7, 0x3b, 0x06, 0x00, 0x00,
 }
 
 func (m *WeightedAddress) Marshal() (dAtA []byte, err error) {
@@ -345,6 +367,34 @@ func (m *Params) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
+	if len(m.WeightedNodeHostsIncentivesReceivers) > 0 {
+		for iNdEx := len(m.WeightedNodeHostsIncentivesReceivers) - 1; iNdEx >= 0; iNdEx-- {
+			{
+				size, err := m.WeightedNodeHostsIncentivesReceivers[iNdEx].MarshalToSizedBuffer(dAtA[:i])
+				if err != nil {
+					return 0, err
+				}
+				i -= size
+				i = encodeVarintParams(dAtA, i, uint64(size))
+			}
+			i--
+			dAtA[i] = 0x22
+		}
+	}
+	if len(m.WeightedNftIncentivesReceivers) > 0 {
+		for iNdEx := len(m.WeightedNftIncentivesReceivers) - 1; iNdEx >= 0; iNdEx-- {
+			{
+				size, err := m.WeightedNftIncentivesReceivers[iNdEx].MarshalToSizedBuffer(dAtA[:i])
+				if err != nil {
+					return 0, err
+				}
+				i -= size
+				i = encodeVarintParams(dAtA, i, uint64(size))
+			}
+			i--
+			dAtA[i] = 0x1a
+		}
+	}
 	if len(m.WeightedDeveloperRewardsReceivers) > 0 {
 		for iNdEx := len(m.WeightedDeveloperRewardsReceivers) - 1; iNdEx >= 0; iNdEx-- {
 			{
@@ -427,6 +477,18 @@ func (m *Params) Size() (n int) {
 	n += 1 + l + sovParams(uint64(l))
 	if len(m.WeightedDeveloperRewardsReceivers) > 0 {
 		for _, e := range m.WeightedDeveloperRewardsReceivers {
+			l = e.Size()
+			n += 1 + l + sovParams(uint64(l))
+		}
+	}
+	if len(m.WeightedNftIncentivesReceivers) > 0 {
+		for _, e := range m.WeightedNftIncentivesReceivers {
+			l = e.Size()
+			n += 1 + l + sovParams(uint64(l))
+		}
+	}
+	if len(m.WeightedNodeHostsIncentivesReceivers) > 0 {
+		for _, e := range m.WeightedNodeHostsIncentivesReceivers {
 			l = e.Size()
 			n += 1 + l + sovParams(uint64(l))
 		}
@@ -869,6 +931,74 @@ func (m *Params) Unmarshal(dAtA []byte) error {
 			}
 			m.WeightedDeveloperRewardsReceivers = append(m.WeightedDeveloperRewardsReceivers, WeightedAddress{})
 			if err := m.WeightedDeveloperRewardsReceivers[len(m.WeightedDeveloperRewardsReceivers)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 3:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field WeightedNftIncentivesReceivers", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthParams
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthParams
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.WeightedNftIncentivesReceivers = append(m.WeightedNftIncentivesReceivers, WeightedAddress{})
+			if err := m.WeightedNftIncentivesReceivers[len(m.WeightedNftIncentivesReceivers)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 4:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field WeightedNodeHostsIncentivesReceivers", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowParams
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthParams
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthParams
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.WeightedNodeHostsIncentivesReceivers = append(m.WeightedNodeHostsIncentivesReceivers, WeightedAddress{})
+			if err := m.WeightedNodeHostsIncentivesReceivers[len(m.WeightedNodeHostsIncentivesReceivers)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex
